requirements.txt
asgiref==3.7.2
astroid==3.0.1
certifi==2024.2.2
charset-normalizer==3.3.2
colorama==0.4.6
dill==0.3.7
distlib==0.3.8
Django==5.0.1
django-filter==23.5
djangorestframework==3.14.0
filelock==3.13.1
idna==3.6
isort==5.12.0
jedi==0.16.0
Markdown==3.5.2
mccabe==0.7.0
numpy==1.26.2
parso==0.8.3
platformdirs==4.0.0
pylint==3.0.2
pytz==2023.3.post1
requests==2.31.0
sqlparse==0.4.4
tomlkit==0.12.3
tzdata==2023.4
urllib3==2.2.0
virtualenv==20.25.0
